+++ NEWS.md
@@ -2,6 +2,8 @@ ClimaLand.jl Release Notes
 ========================
 main
 ----
+- ![][badge-🔥behavioralΔ] use the Sturm et al snow thermal K parameterization, ROSETTA parameters,
+   and new spun up initial conditions. PR [#1772](https://github.com/CliMA/ClimaLand.jl/pull/1766)
 - ![][badge-✨feature] Add selectable snow thermal conductivity parameterizations
   (`JordanSnowConductivityModel`, default, after Jordan 1991; and
   `SturmSnowConductivityModel`, after Sturm et al. 1997) and a piecewise-linear